Wind speed monitoring
The code is used to determine the speed of speed using a reed switch and the customized wind speed model. Reed switch records every count from an electromagnet relay whenever the wind speed model rotates during flow of wind. Greater the speed of wind, the higher the values of reed switch record on the monitor. Speed is calculated by computing all the number of counts after every 15 minutes.